gaussian_splatting: extract chunk-invariant validators into their own TU (decomposition slice 2)
Stacked on the layout-hint extraction (slice 1). Moves the six self-contained
chunk slot-ownership / upload-lifecycle invariant validators out of
gaussian_streaming.cpp's anonymous namespace into a new translation unit
core/streaming_chunk_invariants.{h,cpp} (namespace gs_chunk_invariants):
_release_chunk_slot_if_matches, _chunk_slot_matches_allocator,
_record_streaming_invariant, _validate_pending_upload_chunk_invariant,
_validate_loaded_chunk_invariant, _validate_idle_chunk_invariant.

These six are free helpers over GaussianAtlasAllocator + the streaming
chunk/diagnostics state that reference only each other and their parameters —
a clean partial-class move. The adjacent upload-retirement helpers were
deliberately left in place: four are function templates (instantiated with the
class-nested PendingUploadRetirement) and they depend on non-moved anon
helpers (_subtract_pending_upload_bytes / _release_pending_upload_slot, also
used by out-of-scope code), so moving them is a different, larger change.

gaussian_streaming.cpp includes the new header + adds
`using namespace gs_chunk_invariants;` so all call sites compile unchanged.
The default argument on _chunk_slot_matches_allocator now lives only on the
header declaration (not the definition).

Result: gaussian_streaming.cpp 5592 -> 5459 LOC (6051 at the start of the
decomposition; -592 across slices 1+2).

Risk: R2 (streaming). Evidence: incremental editor build of the combined
slice1+slice2 tree compiles + links clean (scons ... dev_build=yes tests=yes,
exit 0); guards green (run_module_tests.py --guard-only, 108 OK).
